backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
pydsb>=1.0.0
easyocr>=1.6.0
Pillow>=9.5.0
httpx>=0.24.0
python-dotenv>=1.0.0
supabase>=1.0.3
pytest>=7.3.1
pytest-asyncio>=0.21.0
loguru>=0.7.0
numpy>=1.24.0
